Save fin project sans les complexes

+++ include/computorV1.h
@@ -6,7 +6,7 @@
 /*   By: agiordan <marvin@le-101.fr>                +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/03/15 01:59:56 by agiordan     #+#   ##    ##    #+#       */
-/*   Updated: 2019/03/15 08:58:56 by agiordan    ###    #+. /#+    ###.fr     */
+/*   Updated: 2019/03/15 09:32:18 by agiordan    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -15,12 +15,13 @@
 # define COMPUTORV1_H
 
 # include <stdio.h>/////////////////////////////////////////////////////////////
+# include <math.h>/////////////////////////////////////////////////////////////
 # include <unistd.h>
 # include "../libft/libft.h"
 
 typedef struct		s_nome
 {
-	int				coef;
+	double			coef;
 	int				exp;
 	struct s_nome	*next;
 }					t_nome;
@@ -29,7 +30,7 @@ typedef struct		s_parse
 {
 	t_nome			*first;
 	int				iscoef;
-	int				coef;
+	double			coef;
 	int				exp;
 	int				side;
 	int				poly_degree;	
@@ -38,7 +39,7 @@ typedef struct		s_parse
 t_nome				*reduce(t_parse *parse);
 void				resolve(t_parse *parse, t_nome *polynome);
 
-t_nome				*ft_newnome(t_nome **first, int coef, int exp);
+t_nome				*ft_newnome(t_nome **first, double coef, int exp);
 void				ft_delnome(t_nome **first);
 void				ft_print_polynome(t_nome *first);
